Move VirtualAddress and factory into main AsmResolver library.
